[FLINK-28472] Add JmStatus as a group for JobManagerDeploymentStatus count metrics
